// src/app/models/constancia.ts
export class Constancia {
    tipoDeConstancia: string;
    titulo: string;
    solicNumero: string;
    fechaSolic: string;
    nomInstitucion: string;
    locInstitucion: string;
    fechaVisita: string;
    horaVisita: string;
    cantParticipantes: string;
    grado: string;
    orientacion: string;
    nombreResp: string;
    apellidoResp: string;
    mailResp: string;
    telfijoResp: string;
    celResp: string;
    observaciones: string;

}